requirements.txt
pandas>=1.3.3
PyYAML>=5.4.1
requests>=2.26.0
openpyxl>=3.0.7